requirements.txt
fastapi
uvicorn
streamlit
assemblyai
python-dotenv
sqlite-utils
requests
streamlit-audiorecorder
python-multipart
